improve output modules structure
