+++ antarest/output/service.py
@@ -50,7 +50,7 @@ from antarest.launcher.model import LogType
 from antarest.login.utils import get_user_id
 from antarest.matrixstore.service import ISimpleMatrixService
 from antarest.output.dbmodel import Output
-from antarest.output.filestudy.aggregator_management import (
+from antarest.output.filestudy.aggregation import (
     AREA_COL,
     CLUSTER_ID_COL,
     LINK_COL,
@@ -62,7 +62,6 @@ from antarest.output.filestudy.utils import (
     MCIndAreasQueryFile,
     MCIndLinksQueryFile,
     QueryFileType,
-    add_time_index_to_dataframe,
     split_concatenated_columns_from_dataframe,
 )
 from antarest.output.model import (
@@ -574,7 +573,6 @@ class OutputService:
         finally:
             for file_path in file_paths:
                 file_path.unlink(missing_ok=True)
-
         return FileResponse(tmp_file, headers={"Content-Disposition": "inline"}, media_type="application/json")
 
     def delete_output(self, uuid: str, output_name: str) -> None:
@@ -840,7 +838,8 @@ class OutputService:
                 polars_df = polars_df.with_columns(pl.all().cast(pl.Float64))
             df = polars_df.to_pandas()
             if with_index:
-                add_time_index_to_dataframe(df, self.get_output_time_index(study_id, output_id, frequency))
+                matrix_index = self.get_output_time_index(study_id, output_id, frequency)
+                matrix_index.set_as_df_index(df)
             return df
 
         # Checks if the asked couple `variable name` / `output_identifier` exists for the output
