+++ antarest/output/storage/file/abstract_storage.py
@@ -44,10 +44,16 @@ from antarest.core.utils.utils import StopWatch
 from antarest.launcher.adapters.abstractlauncher import SimulationLogs
 from antarest.launcher.model import LogType
 from antarest.matrixstore.in_memory import InMemorySimpleMatrixService
-from antarest.output.filestudy.aggregator_management import AggregatorManager
-from antarest.output.filestudy.file_output_utils import extract_variables_list, parse_output_config
-from antarest.output.filestudy.utils import QueryFileType, get_start_column, parse_output_file_as_pandas_dataframe
-from antarest.output.model import OutputVariablesList
+from antarest.output.filestudy.aggregation import AggregatorManager
+from antarest.output.filestudy.file_output_utils import parse_output_config
+from antarest.output.filestudy.matrixfiles import get_start_column, parse_output_file_as_pandas_dataframe
+from antarest.output.filestudy.utils import (
+    QueryFileType,
+)
+from antarest.output.filestudy.variables import extract_variables_list
+from antarest.output.model import (
+    OutputVariablesList,
+)
 from antarest.output.model.download import MatrixIndex
 from antarest.output.storage.file.repository import FileOutputRepository
 from antarest.output.storage.output_storage import (
