+++ antarest/output/storage/v2/storage.py
@@ -41,10 +41,10 @@ from antarest.launcher.model import LogType
 from antarest.lfs.lfs import ILargeFileStorage
 from antarest.output.filestudy.file_output_utils import (
     extract_output_details,
-    extract_variables_list,
-    find_simulation_log,
 )
+from antarest.output.filestudy.logs import find_simulation_log
 from antarest.output.filestudy.utils import QueryFileType
+from antarest.output.filestudy.variables import extract_variables_list
 from antarest.output.model import OutputVariablesList
 from antarest.output.model.download import MatrixIndex
 from antarest.output.storage.output_storage import (
