+++ antarest/output/storage/v2/variables_storage.py
@@ -33,7 +33,8 @@ from antarest.core.serde.parquet_writer import (
     write_dataframes_stream_parquet,
     yield_dataframes_from_parquet,
 )
-from antarest.output.filestudy.aggregator_management import AggregatorManager
+from antarest.output.filestudy.aggregation import AggregatorManager
+from antarest.output.filestudy.matrixfiles import get_start_column, parse_output_file
 from antarest.output.filestudy.utils import (
     MCYEAR_COL,
     TIME_ID_COL,
@@ -45,9 +46,7 @@ from antarest.output.filestudy.utils import (
     MultipleOutputHeaders,
     QueryFileType,
     get_output_object_type,
-    get_start_column,
     normalize_df_column_names,
-    parse_output_file,
 )
 from antarest.output.utils import find_mode_dir
 from antarest.study.model import MatrixFrequency
